completing the recharging process
