+++ src/main/scala/beam/agentsim/agents/modalbehaviors/DrivesVehicle.scala
@@ -93,7 +93,7 @@ trait DrivesVehicle[T <: DrivingData] extends BeamAgent[T] with HasServices with
                 nextNotifyVehicleResourceIdle = Some(
                   NotifyVehicleResourceIdle(
                     currentVehicleUnderControl,
-                    beamServices.geo.wgs2Utm(currentLeg.travelPath.endPoint),
+                    Some(beamServices.geo.wgs2Utm(currentLeg.travelPath.endPoint)),
                     data.passengerSchedule,
                     theVehicle.getState(),
                     Some(triggerId)
@@ -266,7 +266,7 @@ trait DrivesVehicle[T <: DrivingData] extends BeamAgent[T] with HasServices with
               nextNotifyVehicleResourceIdle = Some(
                 NotifyVehicleResourceIdle(
                   currentVehicleUnderControl,
-                  beamServices.geo.wgs2Utm(updatedBeamLeg.travelPath.endPoint),
+                  Some(beamServices.geo.wgs2Utm(updatedBeamLeg.travelPath.endPoint)),
                   data.passengerSchedule,
                   theVehicle.getState(),
                   _currentTriggerId
@@ -334,43 +334,6 @@ trait DrivesVehicle[T <: DrivingData] extends BeamAgent[T] with HasServices with
   }
 
   when(WaitingToDrive) {
-    case Event(
-        TriggerWithId(EndRefuelTrigger(sessionStart, tick, energyInJoules), triggerId),
-        data
-        ) =>
-      data.currentVehicle.headOption match {
-        case Some(currentVehicleUnderControl) =>
-          val theVehicle = beamServices.vehicles(currentVehicleUnderControl)
-          theVehicle.addFuel(energyInJoules)
-          theVehicle.manager.foreach(
-            _ ! NotifyVehicleResourceIdle(
-              currentVehicleUnderControl,
-              beamServices.geo.wgs2Utm(data.passengerSchedule.schedule.last._1.travelPath.endPoint),
-              data.passengerSchedule,
-              theVehicle.getState(),
-              _currentTriggerId
-            )
-          )
-        case None =>
-          log.debug("currentVehicleUnderControl not found")
-      }
-      stay() replying CompletionNotice(triggerId, Vector())
-    case Event(TriggerWithId(StartRefuelTrigger(tick), triggerId), data) =>
-      data.currentVehicle.headOption match {
-        case Some(currentVehicleUnderControl) =>
-          val theVehicle = beamServices.vehicles(currentVehicleUnderControl)
-          val (sessionDuration, energyDelivered) =
-            theVehicle.refuelingSessionDurationAndEnergyInJoules()
-          stay() replying CompletionNotice(
-            triggerId,
-            Vector(
-              ScheduleTrigger(EndRefuelTrigger(tick + sessionDuration, tick, energyDelivered), self)
-            )
-          )
-        case None =>
-          log.debug("currentVehicleUnderControl not found")
-          stay()
-      }
     case ev @ Event(TriggerWithId(StartLegTrigger(tick, newLeg), triggerId), data) =>
       log.debug("state(DrivesVehicle.WaitingToDrive): {}", ev)
       val triggerToSchedule: Vector[ScheduleTrigger] = data.passengerSchedule
