+++ test/input/beamville/beam.conf
@@ -67,7 +67,7 @@ beam.agentsim.agents.rideHail.defaultCostPerMile=1.25
 beam.agentsim.agents.rideHail.defaultCostPerMinute=0.75
 beam.agentsim.agents.rideHail.vehicleTypeId="BEV"
 beam.agentsim.agents.rideHail.vehicleRangeInMeters=5000.0
-beam.agentsim.agents.rideHail.refuelThresholdInMeters=5000.0
+beam.agentsim.agents.rideHail.refuelThresholdInMeters=4999.0
 # SurgePricing parameters
 beam.agentsim.agents.rideHail.surgePricing.surgeLevelAdaptionStep=0.1
 beam.agentsim.agents.rideHail.surgePricing.minimumSurgeLevel=0.1
